Simplify Perl utf8 handling with "use open ':std', ':encoding(UTF-8)';"
We used a lot of instances of binmode to make STDIN, STDOUT and STDERR work
with utf-8, but we can accomplish that much more directly with just
    use utf8;
    use open ':std', ':encoding(UTF-8)';
at the top of the script.

Also, this syntax enables using the spaceship operator on utf-8 text, both for
STDIN and files listed on the command line:
   while (<>) {
      # process utf8 text in $_
   }
now works correctly, so I no longer need to manually replicate that logic.