fix: make Portage compatible with sh=dash by declaring when we want bash
We routinely use bash-specific syntax in our shell commands and system calls in
Make, Python, Perl, and of course in our shell scripts.

In shell scripts, we make that work with #!/bin/bash
In other languages, we must do the same:
 - make: SHELL := /bin/bash
 - Perl: system("/bin/bash", "-c", $cmd) instead of system($cmd)
	 open(FH, "-|", "/bin/bash", "-c", $cmd) instead of open(FH, "$cmd|")
	 open(FH, "|-", "/bin/bash", "-c", $cmd) instead of open(FH, "|$cmd")
 - Python: add executable="/bin/bash" to calls to subprocess.Popen(),
   .check_output(), .check_call(), .call()
When the commands we write are POSIX-compatible Bourne shell syntax, these
changes are not requred, but when the commands use bash extensions, we must say
so by calling bash.