portage_utils now divides open() between python2 and python3
The Python 2 version is like open in Python 2.7, with no encoding or
newline option. The caller has to deal with that themselves if they want
to. open() only takes care of the magic of files/compressed
files/pipes/stdin/stdout.

The Python 3 version is enriched with features that Python 3's open has:
an encoding parameters, and a newline parameter. Those now have
consistent behaviour, no matter how the file is opened.

In both cases, we strive to have an intuitive implementation, staying
close to what the given version of Python would do.

Exception: in Python 3, when encoding= is not specified, we default to
"utf8", unlike builtins.open(). While less python3-ic because good
Python 3 programming requires providing the encoding on every darn
open() call you make, it makes code more portable, because on Linux
"utf8" ends up being the default most of the time, so testing rarely
flags the problem, while on Windows "utf8" is not typically the default
so we end up with code that's broken on Windows.